New clientlib ver

+++ KalturaClient/src/main/java/com/kaltura/client/types/CuePoint.java
@@ -70,7 +70,7 @@ public abstract class CuePoint extends ObjectBase {
 	}
 
 	private String id;
-	private Integer intId;
+	private Long intId;
 	private CuePointType cuePointType;
 	private CuePointStatus status;
 	private String entryId;
@@ -97,7 +97,7 @@ public abstract class CuePoint extends ObjectBase {
 		return this.id;
 	}
 	// intId:
-	public Integer getIntId(){
+	public Long getIntId(){
 		return this.intId;
 	}
 	// cuePointType:
@@ -260,7 +260,7 @@ public abstract class CuePoint extends ObjectBase {
 
 		// set members values:
 		id = GsonParser.parseString(jsonObject.get("id"));
-		intId = GsonParser.parseInt(jsonObject.get("intId"));
+		intId = GsonParser.parseLong(jsonObject.get("intId"));
 		cuePointType = CuePointType.get(GsonParser.parseString(jsonObject.get("cuePointType")));
 		status = CuePointStatus.get(GsonParser.parseInt(jsonObject.get("status")));
 		entryId = GsonParser.parseString(jsonObject.get("entryId"));
@@ -325,7 +325,7 @@ public abstract class CuePoint extends ObjectBase {
     public CuePoint(Parcel in) {
         super(in);
         this.id = in.readString();
-        this.intId = (Integer)in.readValue(Integer.class.getClassLoader());
+        this.intId = (Long)in.readValue(Long.class.getClassLoader());
         int tmpCuePointType = in.readInt();
         this.cuePointType = tmpCuePointType == -1 ? null : CuePointType.values()[tmpCuePointType];
         int tmpStatus = in.readInt();
